refactor(job-store): improve logging readability and reorganize schema structure
- Valibotバリデーションエラーのログ出力を読みやすく改善
- result.issuesオブジェクトを適切な文字列形式に変換
- HTTPExceptionエラーメッセージを統一化してセキュリティ向上
- スキーマファイルの階層構造を整理して認知負荷を軽減
- 共通スキーマをclient.tsに集約してコードの重複を削除
- jobFetch.tsとjobList/index.tsのインポート構造を簡素化
- 新しいtmp.tsファイルを追加してheadless-crawler統合準備
- typesファイルのインポート構造を整理してモジュール依存関係を明確化
- jobInsert.tsスキーマも整理対象に追加
- 暫定的な整理だが後続開発の制約として機能する構造に変更
