+++ packages/models/src/types/job-store/jobFetch.ts
@@ -1,4 +1,4 @@
 import type { InferOutput } from "valibot";
-import type { JobSchema } from "../../schemas";
+import type { JobSchema } from "../../schemas/job-store/client";
 
 export type Job = InferOutput<typeof JobSchema>;
