+++ packages/models/src/types/job-store/jobList.ts
@@ -1,5 +1,6 @@
 import type { InferOutput } from "valibot";
-import type { decodedNextTokenSchema, JobListSchema } from "../../schemas";
+import type { decodedNextTokenSchema } from "../../schemas";
+import type { JobListSchema } from "../../schemas/job-store/client";
 
 export type JobList = InferOutput<typeof JobListSchema>;
